remove tendermint-abci entirely
