+++ Cargo.toml
@@ -71,8 +71,6 @@ signature                  = "2"
 syn                        = "2"
 tempfile                   = "3"
 tendermint                 = "0.40"
-tendermint-abci            = "0.40"
-tendermint-proto           = "0.40"
 tendermint-rpc             = "0.40"
 test-case                  = "3"
 thiserror                  = "1"
