+++ grug/app/Cargo.toml
@@ -11,25 +11,23 @@ categories    = { workspace = true }
 
 [features]
 default = []
-abci    = ["data-encoding", "tendermint-abci", "tendermint-proto"]
+abci    = ["data-encoding", "tower", "tower-abci"]
 tracing = ["chrono", "dep:tracing"]
 
 [dependencies]
-borsh            = { workspace = true }
-chrono           = { workspace = true, optional = true }
-data-encoding    = { workspace = true, optional = true }
-grug-math        = { workspace = true }
-grug-storage     = { workspace = true }
-grug-types       = { workspace = true }
-ics23            = { workspace = true }
-prost            = { workspace = true }
-serde            = { workspace = true }
-tendermint       = { workspace = true }
-tendermint-abci  = { workspace = true, optional = true }
-tendermint-proto = { workspace = true, optional = true }
-tower            = { workspace = true }
-tower-abci       = { workspace = true }
-thiserror        = { workspace = true }
-tracing          = { workspace = true, optional = true }
+borsh         = { workspace = true }
+chrono        = { workspace = true, optional = true }
+data-encoding = { workspace = true, optional = true }
+grug-math     = { workspace = true }
+grug-storage  = { workspace = true }
+grug-types    = { workspace = true }
+ics23         = { workspace = true }
+prost         = { workspace = true }
+serde         = { workspace = true }
+tendermint    = { workspace = true }
+tower         = { workspace = true, optional = true }
+tower-abci    = { workspace = true, optional = true }
+thiserror     = { workspace = true }
+tracing       = { workspace = true, optional = true }
 
 [dev-dependencies]
