--- grug/app/src/tower.rs
@@ -1,365 +1,0 @@
-use {
-    crate::{App, AppError, AppResult, Db, ProposalPreparer, Vm},
-    grug_math::Inner,
-    grug_types::{
-        Attribute, BlockInfo, Duration, Event, GenericResult, Hash256, Outcome, TxOutcome,
-        GENESIS_BLOCK_HASH,
-    },
-    prost::bytes::Bytes,
-    std::{
-        any::type_name,
-        future::Future,
-        num::NonZeroU32,
-        pin::Pin,
-        task::{Context, Poll},
-    },
-    tendermint::{
-        abci::{
-            request, response, types::ExecTxResult, v0_34::EventAttribute, Code, Event as TmEvent,
-            EventAttribute as TmAttribute,
-        },
-        block::Height,
-        merkle::proof::{ProofOp, ProofOps},
-        v0_38::abci::{Request, Response},
-        AppHash, Hash, Time,
-    },
-    tower::Service,
-    tower_abci::BoxError,
-};
-
-// ----------------------------------- Abci ------------------------------------
-
-impl<DB, VM, PP> Service<Request> for App<DB, VM, PP>
-where
-    DB: Db,
-    VM: Vm + Clone,
-    PP: ProposalPreparer,
-    AppError: From<DB::Error> + From<VM::Error> + From<PP::Error>,
-{
-    type Error = BoxError;
-    type Future =
-        Pin<Box<dyn Future<Output = Result<Self::Response, Self::Error>> + Send + 'static>>;
-    type Response = Response;
-
-    fn poll_ready(&mut self, _: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-        Poll::Ready(Ok(()))
-    }
-
-    fn call(&mut self, req: Request) -> Self::Future {
-        let res: Result<Response, Self::Error> = self
-            .tower_call(req)
-            .map_err(|err| Box::new(err) as BoxError);
-        Box::pin(async move { res })
-    }
-}
-
-impl<DB, VM, PP> App<DB, VM, PP>
-where
-    DB: Db,
-    VM: Vm + Clone,
-    PP: ProposalPreparer,
-    AppError: From<DB::Error> + From<VM::Error> + From<PP::Error>,
-{
-    fn tower_call(&self, req: Request) -> AppResult<Response> {
-        let res = match req {
-            // ----------------------- Hanlded requests ------------------------
-            Request::CheckTx(req) => Response::CheckTx(self.tower_check_tx(req)?),
-            Request::Commit => Response::Commit(self.tower_commit()?),
-            Request::FinalizeBlock(req) => Response::FinalizeBlock(self.tower_finalize_block(req)?),
-            Request::Info(..) => Response::Info(self.tower_info()?),
-            Request::InitChain(req) => Response::InitChain(self.tower_init_chain(req)?),
-            Request::PrepareProposal(req) => {
-                Response::PrepareProposal(self.tower_prepare_proposal(req)?)
-            },
-            Request::Query(req) => Response::Query(self.tower_query(req)?),
-
-            // ----------------------- Unhandled requests ----------------------
-            Request::ApplySnapshotChunk(_) => Response::ApplySnapshotChunk(Default::default()),
-            Request::Echo(_) => Response::Echo(Default::default()),
-            Request::ExtendVote(_) => Response::ExtendVote(response::ExtendVote {
-                vote_extension: Bytes::default(),
-            }),
-            Request::Flush => Response::Flush,
-            Request::ListSnapshots => Response::ListSnapshots(Default::default()),
-            Request::LoadSnapshotChunk(_) => Response::LoadSnapshotChunk(Default::default()),
-            Request::OfferSnapshot(_) => Response::OfferSnapshot(Default::default()),
-            Request::ProcessProposal(_) => {
-                Response::ProcessProposal(response::ProcessProposal::Accept)
-            },
-            Request::VerifyVoteExtension(_) => {
-                Response::VerifyVoteExtension(response::VerifyVoteExtension::Accept)
-            },
-        };
-
-        Ok(res)
-    }
-
-    fn tower_check_tx(&self, req: request::CheckTx) -> AppResult<response::CheckTx> {
-        let res = match self.do_check_tx_raw(&req.tx) {
-            Ok(Outcome {
-                result: GenericResult::Ok(events),
-                gas_limit,
-                ..
-            }) => response::CheckTx {
-                code: Code::Ok,
-                events: into_tm_events(events),
-                gas_wanted: gas_limit.unwrap() as i64,
-                // Note: Return `Outcome::gas_limited` instead of `gas_used here.
-                // This is because in `CheckTx` we don't run the entire tx, just
-                // the authentication part. As such, the gas consumption is
-                // underestimated. Instead, the tx gas limit represents the max
-                // amount of gas this tx can possibly consume.
-                gas_used: gas_limit.unwrap() as i64,
-                ..Default::default()
-            },
-            Ok(Outcome {
-                result: GenericResult::Err(err),
-                gas_limit,
-                ..
-            }) => response::CheckTx {
-                code: code_error(1),
-                codespace: "tx".into(),
-                log: err,
-                gas_wanted: gas_limit.unwrap() as i64,
-                gas_used: gas_limit.unwrap() as i64,
-                ..Default::default()
-            },
-            Err(err) => response::CheckTx {
-                code: code_error(1),
-                codespace: "simulate".into(),
-                log: err.to_string(),
-                ..Default::default()
-            },
-        };
-
-        Ok(res)
-    }
-
-    fn tower_commit(&self) -> AppResult<response::Commit> {
-        match self.do_commit() {
-            Ok(()) => Ok(response::Commit {
-                // This field is ignored since CometBFT 0.38.
-                // TODO: Can we omit this?????
-                data: Default::default(),
-                // TODO: what this means??
-                retain_height: Height::default(),
-            }),
-            Err(err) => panic!("failed to commit: {err}"),
-        }
-    }
-
-    fn tower_finalize_block(
-        &self,
-        req: request::FinalizeBlock,
-    ) -> AppResult<response::FinalizeBlock> {
-        let block = from_tm_block(req.height.value(), req.time, Some(req.hash));
-
-        match self.do_finalize_block_raw(block, &req.txs) {
-            Ok(outcome) => {
-                // In Cosmos SDK, this refers to the Begin/EndBlocker events.
-                // For us, this is the cronjob events.
-                // Note that failed cronjobs are ignored (not included in `ResponseFinalizeBlock`).
-                let events = outcome
-                    .cron_outcomes
-                    .into_iter()
-                    .filter_map(|outcome| outcome.result.ok().map(into_tm_events))
-                    .flatten()
-                    .collect();
-
-                let tx_results = outcome
-                    .tx_outcomes
-                    .into_iter()
-                    .map(into_tm_tx_result)
-                    .collect();
-
-                Ok(response::FinalizeBlock {
-                    app_hash: into_tm_app_hash(outcome.app_hash),
-                    events,
-                    tx_results,
-                    // We haven't implemented any mechanism to alter the
-                    // validator set or consensus params yet.
-                    validator_updates: vec![],
-                    consensus_param_updates: None,
-                })
-            },
-            Err(err) => panic!("failed to finalize block: {err}"),
-        }
-    }
-
-    fn tower_info(&self) -> AppResult<response::Info> {
-        let (last_block_height, last_block_version) = self.do_info()?;
-        Ok(response::Info {
-            data: env!("CARGO_PKG_NAME").into(),
-            version: env!("CARGO_PKG_VERSION").into(),
-            app_version: 1,
-            last_block_height: last_block_height
-                .try_into()
-                .expect("block height exceeds i64"),
-            last_block_app_hash: into_tm_app_hash(last_block_version),
-        })
-    }
-
-    fn tower_init_chain(&self, req: request::InitChain) -> AppResult<response::InitChain> {
-        let block = from_tm_block(0, req.time, None);
-
-        match self.do_init_chain_raw(req.chain_id, block, &req.app_state_bytes) {
-            Ok(app_hash) => Ok(response::InitChain {
-                consensus_params: Some(req.consensus_params),
-                validators: req.validators,
-                app_hash: into_tm_app_hash(app_hash),
-            }),
-            Err(err) => panic!("failed to init chain: {err}"),
-        }
-    }
-
-    fn tower_prepare_proposal(
-        &self,
-        req: request::PrepareProposal,
-    ) -> AppResult<response::PrepareProposal> {
-        let max_tx_bytes = req.max_tx_bytes.try_into().unwrap_or(0);
-        let txs = self.do_prepare_proposal(req.txs.clone(), max_tx_bytes);
-
-        Ok(response::PrepareProposal { txs })
-    }
-
-    fn tower_query(&self, req: request::Query) -> AppResult<response::Query> {
-        let res = match req.path.as_str() {
-            "/app" => match self.do_query_app_raw(&req.data, req.height.value(), req.prove) {
-                Ok(res) => response::Query {
-                    code: Code::Ok,
-                    value: res.into(),
-                    ..Default::default()
-                },
-                Err(err) => response::Query {
-                    code: Code::Ok,
-                    codespace: "app".into(),
-                    log: err.to_string(),
-                    ..Default::default()
-                },
-            },
-            "/simulate" => match self.do_simulate_raw(&req.data, req.height.value(), req.prove) {
-                Ok(outcome) => response::Query {
-                    code: Code::Ok,
-                    value: outcome.into(),
-                    ..Default::default()
-                },
-                Err(err) => response::Query {
-                    code: code_error(1),
-                    codespace: "simulate".into(),
-                    log: err.to_string(),
-                    ..Default::default()
-                },
-            },
-            "/store" => match self.do_query_store(&req.data, req.height.value(), req.prove) {
-                Ok((value, proof)) => {
-                    let proof = proof.map(|proof| ProofOps {
-                        ops: vec![ProofOp {
-                            field_type: type_name::<DB::Proof>().into(),
-                            key: req.data.into(),
-                            data: proof,
-                        }],
-                    });
-                    response::Query {
-                        code: Code::Ok,
-                        value: value.unwrap_or_default().into(),
-                        height: req.height,
-                        proof,
-                        ..Default::default()
-                    }
-                },
-                Err(err) => response::Query {
-                    code: code_error(1),
-                    codespace: "store".into(),
-                    log: err.to_string(),
-                    ..Default::default()
-                },
-            },
-            unknown => response::Query {
-                code: code_error(1),
-                codespace: "app".into(),
-                log: format!("unknown path `{unknown}`; must be `/app`, `/simulate`, or `/store`"),
-                ..Default::default()
-            },
-        };
-
-        Ok(res)
-    }
-}
-
-// ----------------------------------- Utils -----------------------------------
-
-fn from_tm_block(height: u64, time: Time, hash: Option<Hash>) -> BlockInfo {
-    BlockInfo {
-        height,
-        timestamp: Duration::from_nanos(time.unix_timestamp_nanos() as u128),
-        hash: hash.map(from_tm_hash).unwrap_or(GENESIS_BLOCK_HASH),
-    }
-}
-
-fn from_tm_hash(bytes: Hash) -> Hash256 {
-    match bytes {
-        Hash::Sha256(hash) => Hash256::from_inner(hash),
-        Hash::None => panic!("unexpected empty hash"),
-    }
-}
-
-fn into_tm_tx_result(outcome: TxOutcome) -> ExecTxResult {
-    match outcome.result {
-        GenericResult::Ok(_) => ExecTxResult {
-            code: Code::Ok,
-            gas_wanted: outcome.gas_limit as i64,
-            gas_used: outcome.gas_used as i64,
-            events: into_tm_events(outcome.events),
-            ..Default::default()
-        },
-        GenericResult::Err(err) => ExecTxResult {
-            code: code_error(1),
-            codespace: "tx".to_string(),
-            log: err,
-            gas_wanted: outcome.gas_limit as i64,
-            gas_used: outcome.gas_used as i64,
-            events: into_tm_events(outcome.events),
-            ..Default::default()
-        },
-    }
-}
-
-fn into_tm_events<I>(events: I) -> Vec<TmEvent>
-where
-    I: IntoIterator<Item = Event>,
-{
-    events.into_iter().map(into_tm_event).collect()
-}
-
-fn into_tm_event(event: Event) -> TmEvent {
-    TmEvent {
-        kind: event.r#type,
-        attributes: into_tm_attributes(event.attributes),
-    }
-}
-
-fn into_tm_attributes<I>(attrs: I) -> Vec<TmAttribute>
-where
-    I: IntoIterator<Item = Attribute>,
-{
-    attrs.into_iter().map(into_tm_attribute).collect()
-}
-
-fn into_tm_attribute(attr: Attribute) -> TmAttribute {
-    // V037 is not exported from the `tendermint` crate.
-    // IDK how to import it.
-    TmAttribute::V034(EventAttribute {
-        key: attr.key.as_bytes().to_vec(),
-        value: attr.value.as_bytes().to_vec(),
-        index: true,
-    })
-}
-
-fn into_tm_app_hash(hash: Hash256) -> AppHash {
-    hash.into_inner().to_vec().try_into().unwrap()
-}
-
-/// Be sure to pass a non-zero index
-fn code_error(index: u32) -> Code {
-    Code::Err(unsafe { NonZeroU32::new_unchecked(index) })
-}
